[WIP] Don't consider URL starting with 'username@' as secure
This has been implemented [1] but without any reference to the actual
usecase. As it turned out, such URL scheme is also known as 'scp-like'
and used, for instance, by Git [2]:

    > git clone user@server:project.git

is a shortcut for

    > git clone ssh://user@server/project.git

Not sure, but seems it's not an 'el-get-insecure-check responsibility to
handle such provider-specific cases. Implementation of the specific
:type should create a valid URL.

[1] acdcb6e5b2e86e401f3e1738723a40a5f2b654c5
[2] manual for 'git fetch'

As there is no formal definition of SCP-like URLs, guess work follows.
Based on 3.2.2. Server-based Naming Authority from RFC 2396 [3].

<user>@<host>:<path>

userinfo = *( unreserved | escaped | ";" | ":" | "&" | "=" | "+" | "$" | "," )

unreserved  = alphanum | mark
mark = "-" | "_" | "." | "!" | "~" | "*" | "'" | "(" | ")"

[3] https://www.ietf.org/rfc/rfc2396.txt